+++ src/test/java/roomescape/repository/JpaPaymentRepositoryTest.java
@@ -2,6 +2,7 @@ package roomescape.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,8 @@ class JpaPaymentRepositoryTest {
     @Test
     @DisplayName("Payment 를 잘 저장하는지 확인한다.")
     void save() {
-        Payment saved = paymentRepository.save(new Payment(null, "paymentKey1", "WTEST12345", 1000L));
+        Payment saved = paymentRepository.save(
+                new Payment(null, "paymentKey1", "WTEST12345", BigDecimal.valueOf(1000L)));
 
         assertThat(saved.getId()).isNotNull();
     }
